# Conflicts:
#	BlazorWASM-AppLogic/Application/Application.csproj
#	BlazorWASM-AppLogic/Application/GrpcClients/PostGrpcClient.cs
#	BlazorWASM-AppLogic/Application/Logic/PostLogic.cs
#	BlazorWASM-AppLogic/Application/Protos/Post.proto
#	BlazorWASM-AppLogic/BlazorWASM/Pages/CreatePost.razor
#	BlazorWASM-AppLogic/BlazorWASM/Pages/PostDetails.razor
#	BlazorWASM-AppLogic/BlazorWASM/Pages/ViewPosts.razor
#	BlazorWASM-AppLogic/BlazorWASM/Program.cs
#	BlazorWASM-AppLogic/Domain/DTOs/PostCreationDto.cs
#	BlazorWASM-AppLogic/Domain/Models/Post.cs
#	DataPersistence/src/main/java/sep3/project/daos/impl/PostDatabase.java
#	DataPersistence/src/main/java/sep3/project/daos/interfaces/PostPersistence.java
#	DataPersistence/src/main/java/sep3/project/protobuf/Post.java
#	DataPersistence/src/main/java/sep3/project/protobuf/PostData.java
#	DataPersistence/src/main/java/sep3/project/protobuf/PostDataOrBuilder.java
#	DataPersistence/src/main/java/sep3/project/protobuf/RequestCreatePost.java
#	DataPersistence/src/main/java/sep3/project/protobuf/RequestCreatePostOrBuilder.java
#	DataPersistence/src/main/java/sep3/project/protobuf/User.java
#	DataPersistence/src/main/java/sep3/project/protobuf/UserServiceGrpc.java
#	DataPersistence/src/main/java/sep3/project/services/PostImpl.java
#	DataPersistence/src/main/proto/post.proto

+++ BlazorWASM-AppLogic/Domain/Models/Coordinate.cs
@@ -2,12 +2,13 @@
 
 public class Coordinate
 {
-    public float latitude{ get; set; }
-    public float longitude{ get; set; }
+    public double latitude{ get; set; }
+    public double longitude{ get; set; }
 
-    public Coordinate(float latitude, float longitude)
+    public Coordinate(double latitude, double longitude)
     {
         this.latitude = latitude;
         this.longitude = longitude;
     }
+
 }
